requirements-dev.txt
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.28.0
mongomock>=4.2.0
pytest-cov>=5.0.0
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.18.2
